backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
email-validator==2.1.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx>=0.28.0
qrcode[pil]==7.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
requests==2.31.0
aptos-sdk==0.11.0
pyjwt==2.8.0
slowapi==0.1.9
redis==5.0.1
cryptography==41.0.7
gunicorn==21.2.0
email-validator==2.1.0
